requirements.txt
Django==2.0.7
django-debug-toolbar==1.9.1
django-summernote==0.8.8.7
mysqlclient==1.3.13
pytz==2018.5
sqlparse==0.2.4
